.env API

+++ src/app/api/delete.ts
@@ -1,5 +1,16 @@
-export const deleteCallApi = async(url : string, data : any, accessToken? : string) => {
+export const deleteCallApi = async (
+  endpoint: string,
+  data: any,
+  accessToken?: string,
+) => {
   try {
+    const baseUrl = process.env.NEXT_PUBLIC_API_BASE_URL;
+    if (!baseUrl) {
+      console.error('API_BASE_URL is not defined in .env.local');
+      throw new Error('API_BASE_URL is not defined in .env.local');
+    }
+
+    const url = `${baseUrl}${endpoint}`;
     const response = await fetch(url, {
       method: 'DELETE',
       headers: {
@@ -8,7 +19,12 @@ export const deleteCallApi = async(url : string, data : any, accessToken? : stri
       },
       body: JSON.stringify(data),
     });
-   return response.json
+
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+
+    return await response.json();
   } catch (error) {
     console.error('Erreur lors de la connexion au serveur', error);
     return {
